+++ orix/tests/test_object3d.py
@@ -16,8 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with orix.  If not, see <http://www.gnu.org/licenses/>.
 
-import pytest
 import numpy as np
+import pytest
 
 from orix.base import DimensionError, Object3d, check